FE-2273 - RadioButton sets default value by defaultChecked Exported common RadioGroup behaviour to RadioButtonMapper component
