+++ src/__experimental__/components/radio-button/radio-button-group.component.js
@@ -1,64 +1,17 @@
-import React, { useState, useCallback, useMemo } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
+
 import tagComponent from '../../../utils/helpers/tags';
 import RadioButtonFieldsetStyle from './radio-button-fieldset.style';
+import RadioButtonMapper from './radio-button-mapper.component';
 import withValidation from '../../../components/validations/with-validation.hoc';
 
 const RadioButtonGroup = (props) => {
   const {
-    children,
-    name,
-    legend,
-    hasError,
-    hasWarning,
-    hasInfo,
-    onBlur,
-    onChange,
-    value,
-    tooltipMessage
+    children, name, legend, hasError, hasWarning, hasInfo, onBlur, onChange, value, tooltipMessage
   } = props;
+
   const groupLabelId = `${name}-label`;
-
-  const anyChecked = useMemo(() => {
-    let result = false;
-    React.Children.map(children, (child) => {
-      if (Object.prototype.hasOwnProperty.call(child.props, 'checked')) {
-        result = true;
-      }
-    });
-    return result;
-  }, [children]);
-
-  const isControlled = value !== undefined;
-
-  const [checkedValue, setCheckedValue] = useState(false);
-  const onChangeProp = useCallback((e) => {
-    onChange(e);
-    if (!isControlled) {
-      setCheckedValue(e.target.value);
-    }
-  }, [onChange, setCheckedValue, isControlled]);
-
-  const buttons = React.Children.map(children, (child) => {
-    let checked;
-    if (isControlled) {
-      // The user is controlling the input via the value prop
-      checked = value === child.props.value;
-    } else if (!checkedValue && anyChecked) {
-      // Uncontrolled and the user has not made a selection, but at least one has a checked prop
-      checked = child.props.checked || false;
-    } else {
-      // Uncontrolled, existing selection or none marked as checked
-      checked = checkedValue === child.props.value;
-    }
-
-    return React.cloneElement(child, {
-      checked,
-      name,
-      onBlur,
-      onChange: onChangeProp
-    });
-  });
 
   return (
     <RadioButtonFieldsetStyle
@@ -71,7 +24,14 @@ const RadioButtonGroup = (props) => {
       tooltipMessage={ tooltipMessage }
       { ...tagComponent('radiogroup', props) }
     >
-      {buttons}
+      <RadioButtonMapper
+        name={ name }
+        onBlur={ onBlur }
+        onChange={ onChange }
+        value={ value }
+      >
+        {children}
+      </RadioButtonMapper>
     </RadioButtonFieldsetStyle>
   );
 };
